refactor(misc): co-locate phase classes in modules + RunOutputs convenience views (#239)
